improve scaling accuracy, prevent overflow

+++ src/atrac/atrac1_scale.h
@@ -14,7 +14,7 @@ struct TScaledBlock {
 };
 
 class TScaler : public TAtrac1Data {
-    static std::map<uint32_t, uint8_t>ScaleIndex;
+    static std::map<double, uint8_t>ScaleIndex;
 public:
     TScaler();
     std::vector<TScaledBlock> Scale(const std::vector<double>& specs, const TBlockSize& blockSize);
